# Text performance tweaks