… examples/feature_demo/test_big.py
… tests/geometries/test_text_geometry.py
… tests/utils/test_text_shaping.py