@Korijn 2025-02-28
Just out of curiosity: are you using a profiler?